+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Projcet: 100 stories
+- Project: hiyori
 ### Changed
 - Updated: storybuilder (v0.1.0)