+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+
+## [0.0.3] - 2019-04-20
+### Added
+- Project: 100 characters
 - Projcet: 100 stories
 - Project: hiyori
 ### Changed
@@ -27,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Production: slime love story
 - Production: kujo saeko
 
-[Unreleased]: https://github.com/nagisc007/prj_estar/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/nagisc007/prj_estar/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/nagisc007/prj_estar/releases/v0.0.3
 [0.0.2]: https://github.com/nagisc007/prj_estar/releases/v0.0.2
 [0.0.1]: https://github.com/nagisc007/prj_estar/releases/v0.0.1